Android app to alert when exceeding set speeds
The various files form part of an Android app for alerting when a set speed is being exceeded. The layout is designed for both Landscape and in Portrait modes. You will see a set of icons representing set speeds, when one of these icons is pressed, the speed to that number is set, in kilometres per hour. When that speed is being exceeded, an MP3 file, which contains a car horn beeping, will be played until the speed goes back below the set speed.